backend/requirements.txt
firebase-admin
google-cloud-firestore
google-cloud-bigquery
python-dotenv
shapely
google-generativeai
imagehash
Pillow
sentence-transformers
scikit-learn
numpy